Add greeting component
