+++ src/App.js
@@ -1,28 +1,12 @@
-import { useEffect } from 'react';
 import './App.css';
-import { useSelector, useDispatch } from 'react-redux';
-import { fetchGreeting } from './redux/slices/greetingsSlice';
+import Greeting from './components/Greeting';
 
 function App() {
-  const dispatch = useDispatch();
-  const greeting = useSelector((state) => state.greeting);
-  const loading = useSelector((state) => state.loading);
-  const error = useSelector((state) => state.error);
-  console.log(greeting);
-
-  useEffect(() => {
-    dispatch(fetchGreeting());
-  }, []);
-
-  if (loading) {
-    return <div className="App">Loading...</div>;
-  }
-
-  if (error) {
-    return <div className="App">{error}</div>;
-  }
-
-  return <div className="App">{greeting.greeting}</div>;
+  return (
+    <div className="App">
+      <Greeting />
+    </div>
+  );
 }
 
 export default App;
