+++ src/components/Greeting.jsx
@@ -1,0 +1,25 @@
+import React, { useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { fetchGreeting } from '../redux/slices/greetingsSlice';
+
+const Greeting = () => {
+  const dispatch = useDispatch();
+  const greeting = useSelector((state) => state.greeting);
+  const loading = useSelector((state) => state.loading);
+  const error = useSelector((state) => state.error);
+
+  useEffect(() => {
+    dispatch(fetchGreeting());
+  }, []);
+
+  if (loading) {
+    return <div className="App">Loading...</div>;
+  }
+
+  if (error) {
+    return <div className="App">{error}</div>;
+  }
+  return <div className="App">{greeting.greeting}</div>;
+};
+
+export default Greeting;
